Added example; check content; function renaming.
Version 1.9.0

Added glazyCheckContent() function that works similar to
glazyAttributeCheck(). Content that is checked by this and passed to
the glazy- functions that doesn’t exist will cause element to not be
displayed.

Renamed:
glazyPrepareContentWithSpacing() to glazyPrepareContentJoinedBy()
glazyPrepareContentJoined() to glazyPrepareContent()
glazyPrepareContentWithSpacing() to glazyPrepareContentJoinedBy()
glazyPrepareContentWithLineBreaks() to
glazyPrepareContentJoinedByLineBreaks()

Tags html, head, body are all displayed with \n after.

Added example files for previewing and checking output.